requirements.txt
fastapi
uvicorn
aiohttp
python-multipart
jinja2
python-dotenv
requests
urllib3